fix: don't reset amounts on switch tokens
