+++ src/modules/trade/hooks/useIsWrapOrUnwrap.ts
@@ -1,13 +1,11 @@
 import { useMemo } from 'react'
 
-import { NATIVE_CURRENCY_BUY_TOKEN } from 'legacy/constants'
-import { WRAPPED_NATIVE_CURRENCY } from 'legacy/constants/tokens'
 import { supportedChainId } from 'legacy/utils/supportedChainId'
 
 import { useTradeState } from 'modules/trade/hooks/useTradeState'
 import { useWalletInfo } from 'modules/wallet'
 
-import { doesTokenMatchSymbolOrAddress } from 'utils/doesTokenMatchSymbolOrAddress'
+import { getIsWrapOrUnwrap } from 'utils/getIsWrapOrUnwrap'
 
 export function useIsWrapOrUnwrap(): boolean {
   const { chainId } = useWalletInfo()
@@ -17,17 +15,6 @@ export function useIsWrapOrUnwrap(): boolean {
   return useMemo(() => {
     if (!chainId || !supportedChainId(chainId)) return false
 
-    if (!inputCurrencyId || !outputCurrencyId) return false
-
-    const nativeToken = NATIVE_CURRENCY_BUY_TOKEN[chainId]
-    const wrappedToken = WRAPPED_NATIVE_CURRENCY[chainId]
-
-    const isNativeIn = doesTokenMatchSymbolOrAddress(nativeToken, inputCurrencyId)
-    const isNativeOut = doesTokenMatchSymbolOrAddress(nativeToken, outputCurrencyId)
-
-    const isWrappedIn = doesTokenMatchSymbolOrAddress(wrappedToken, inputCurrencyId)
-    const isWrappedOut = doesTokenMatchSymbolOrAddress(wrappedToken, outputCurrencyId)
-
-    return (isNativeIn && isWrappedOut) || (isNativeOut && isWrappedIn)
+    return getIsWrapOrUnwrap(chainId, inputCurrencyId, outputCurrencyId)
   }, [chainId, inputCurrencyId, outputCurrencyId])
 }
